чтение координат аэропортов через geopify API

+++ src/main/java/com/aviaservice/airportsSystem/dto/AirportDto.java
@@ -19,8 +19,8 @@ public class AirportDto extends VersionedDto {
 
     private String name;
 
-    private BigDecimal lon;
+    private double lon;
 
-    private BigDecimal lat;
+    private double lat;
 
 }
